MISRA 10.4 rule

+++ src/api/api.c
@@ -33,7 +33,7 @@ z_keyexpr_t z_keyexpr(const char *name) { return _z_rname(name); }
 char *z_keyexpr_to_string(z_keyexpr_t keyexpr) {
     if (keyexpr._id != Z_RESOURCE_ID_NONE) return NULL;
 
-    size_t ke_len = strlen(keyexpr._suffix) + 1;
+    size_t ke_len = strlen(keyexpr._suffix) + (size_t)1;
     char *ret = (char *)z_malloc(ke_len);
     strncpy(ret, keyexpr._suffix, ke_len);
