+++ src/link/unicast/tcp.c
@@ -31,7 +31,7 @@ char *_z_parse_port_segment_tcp(char *address) {
     char *p_end = &address[strlen(address)];
 
     int len = p_end - p_start;
-    char *port = (char *)z_malloc((len + 1) * sizeof(char));
+    char *port = (char *)z_malloc((len + 1) * (int)sizeof(char));
     strncpy(port, p_start, len);
     port[len] = '\0';
 
@@ -47,7 +47,7 @@ char *_z_parse_address_segment_tcp(char *address) {
         p_start++;
         p_end--;
         int len = p_end - p_start;
-        char *ip6_addr = (char *)z_malloc((len + 1) * sizeof(char));
+        char *ip6_addr = (char *)z_malloc((len + 1) * (int)sizeof(char));
         strncpy(ip6_addr, p_start, len);
         ip6_addr[len] = '\0';
 
@@ -56,7 +56,7 @@ char *_z_parse_address_segment_tcp(char *address) {
     // IPv4
     else {
         int len = p_end - p_start;
-        char *ip4_addr_or_domain = (char *)z_malloc((len + 1) * sizeof(char));
+        char *ip4_addr_or_domain = (char *)z_malloc((len + 1) * (int)sizeof(char));
         strncpy(ip4_addr_or_domain, p_start, len);
         ip4_addr_or_domain[len] = '\0';
