+++ src/protocol/codec.c
@@ -147,7 +147,7 @@ _z_str_result_t _z_str_decode(_z_zbuf_t *zbf) {
     }
 
     // Allocate space for the string terminator
-    char *s = (char *)z_malloc(len + 1);
+    char *s = (char *)z_malloc(len + (size_t)1);
     s[len] = '\0';
     _z_zbuf_read_bytes(zbf, (uint8_t *)s, 0, len);
     r._value._str = s;
