+++ src/system/espidf/network.c
@@ -56,8 +56,8 @@ void *_z_open_tcp(void *arg, uint32_t tout) {
     if (sock < 0) goto _Z_OPEN_TCP_ERROR_1;
 
     z_time_t tv;
-    tv.tv_sec = tout / 1000;
-    tv.tv_usec = (tout % 1000) * 1000;
+    tv.tv_sec = tout / (uint32_t)1000;
+    tv.tv_usec = (tout % (uint32_t)1000) * (uint32_t)1000;
     if (setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, (char *)&tv, sizeof(tv)) < 0) goto _Z_OPEN_TCP_ERROR_2;
 
     int flags = 1;
@@ -124,7 +124,7 @@ size_t _z_read_exact_tcp(void *sock_arg, uint8_t *ptr, size_t len) {
 
         n -= rb;
         ptr = ptr + (len - n);
-    } while (n > 0);
+    } while (n > (size_t)0);
 
     return len;
 }
@@ -168,8 +168,8 @@ void *_z_open_udp_unicast(void *arg, uint32_t tout) {
     if (sock < 0) goto _Z_OPEN_UDP_UNICAST_ERROR_1;
 
     z_time_t tv;
-    tv.tv_sec = tout / 1000;
-    tv.tv_usec = (tout % 1000) * 1000;
+    tv.tv_sec = tout / (uint32_t)1000;
+    tv.tv_usec = (tout % (uint32_t)1000) * (uint32_t)1000;
     if (setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, (char *)&tv, sizeof(tv)) < 0) goto _Z_OPEN_UDP_UNICAST_ERROR_2;
 
     ret->_fd = sock;
@@ -220,7 +220,7 @@ size_t _z_read_exact_udp_unicast(void *sock_arg, uint8_t *ptr, size_t len) {
 
         n -= rb;
         ptr = ptr + (len - n);
-    } while (n > 0);
+    } while (n > (size_t)0);
 
     return len;
 }
@@ -267,8 +267,8 @@ void *_z_open_udp_multicast(void *arg_1, void **arg_2, uint32_t tout, const char
     if (sock < 0) goto _Z_OPEN_UDP_MULTICAST_ERROR_2;
 
     z_time_t tv;
-    tv.tv_sec = tout / 1000;
-    tv.tv_usec = (tout % 1000) * 1000;
+    tv.tv_sec = tout / (uint32_t)1000;
+    tv.tv_usec = (tout % (uint32_t)1000) * (uint32_t)1000;
     if (setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, (char *)&tv, sizeof(tv)) < 0) goto _Z_OPEN_UDP_MULTICAST_ERROR_3;
 
     if (bind(sock, lsockaddr, addrlen) < 0) goto _Z_OPEN_UDP_MULTICAST_ERROR_3;
@@ -344,8 +344,8 @@ void *_z_listen_udp_multicast(void *arg, uint32_t tout, const char *iface) {
     if (sock < 0) goto _Z_LISTEN_UDP_MULTICAST_ERROR_1;
 
     z_time_t tv;
-    tv.tv_sec = tout / 1000;
-    tv.tv_usec = (tout % 1000) * 1000;
+    tv.tv_sec = tout / (uint32_t)1000;
+    tv.tv_usec = (tout % (uint32_t)1000) * (uint32_t)1000;
     if (setsockopt(sock, SOL_SOCKET, SO_RCVTIMEO, (char *)&tv, sizeof(tv)) < 0) goto _Z_LISTEN_UDP_MULTICAST_ERROR_2;
 
     int optflag = 1;
@@ -467,7 +467,7 @@ size_t _z_read_exact_udp_multicast(void *sock_arg, uint8_t *ptr, size_t len, voi
 
         n -= rb;
         ptr = ptr + (len - n);
-    } while (n > 0);
+    } while (n > (size_t)0);
 
     return len;
 }
